+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - **[BREAKING CHANGE]** Remove Microdata from `TripCard`
 - **[BREAKING CHANGE]** Use new `Itinerary` in `TripCard`
+- **[BREAKING CHANGE]** `Caption` is now private (the component, not the typography element)
+- **[FIX]** Fix broken docs for dividers
 [...]
 
 # v52.0.0 (03/03/2021)